+++ DesignToolbox/DesignToolbox/Pages/Components/Checkbox/CheckboxItem/CheckboxItemIndeterminateConfiguration.swift
@@ -99,7 +99,7 @@ final class CheckboxItemIndeterminateConfigurationModel: ComponentConfiguration 
     // swiftlint:enable line_length
 
     private var disableCode: String {
-        ".disable(\(enabled ? "false" : "true"))"
+        ".disabled(\(enabled ? "false" : "true"))"
     }
 
     private var helperTextPattern: String {