+++ DesignToolbox/DesignToolbox/Pages/Components/Radio/Radio/RadioConfiguration.swift
@@ -54,7 +54,7 @@ final class RadioConfigurationModel: ComponentConfiguration {
     }
 
     private var disableCode: String {
-        ".disable(\(enabled ? "false" : "true"))"
+        ".disabled(\(enabled ? "false" : "true"))"
     }
 
     private var isErrorPattern: String {