+++ DesignToolbox/DesignToolbox/Pages/Components/Switch/Switchitem/SwitchItemConfiguration.swift
@@ -14,10 +14,10 @@
 import OUDSComponents
 import SwiftUI
 
-// MARK: - Radio Item Configuration Model
+// MARK: - Switch Item Configuration Model
 
-/// The model shared between `RadioItemConfiguration` view and `RadioItemPage` view.
-final class RadioItemConfigurationModel: ComponentConfiguration {
+/// The model shared between `SwitchItemConfiguration` view and `SwitchItemPage` view.
+final class SwitchItemConfigurationModel: ComponentConfiguration {
 
     // MARK: - Properties
 
@@ -26,10 +26,6 @@ final class RadioItemConfigurationModel: ComponentConfiguration {
     }
 
     @Published var selection: Bool {
-        didSet { updateCode() }
-    }
-
-    @Published var additionalLabelText: Bool {
         didSet { updateCode() }
     }
 
@@ -53,10 +49,6 @@ final class RadioItemConfigurationModel: ComponentConfiguration {
         didSet { updateCode() }
     }
 
-    @Published var outlined: Bool {
-        didSet { updateCode() }
-    }
-
     @Published var divider: Bool {
         didSet { updateCode() }
     }
@@ -66,10 +58,6 @@ final class RadioItemConfigurationModel: ComponentConfiguration {
     }
 
     @Published var labelTextContent: String {
-        didSet { updateCode() }
-    }
-
-    @Published var additionalLabelTextContent: String {
         didSet { updateCode() }
     }
 
@@ -83,17 +71,14 @@ final class RadioItemConfigurationModel: ComponentConfiguration {
         selection = true
         isError = false
         isReadOnly = false
-        outlined = false
         enabled = true
-        additionalLabelText = true
         helperText = true
         icon = true
         flipIcon = false
         layoutOrientation = .default
         divider = true
-        labelTextContent = String(localized: "app_components_radio_label_text")
-        additionalLabelTextContent = String(localized: "app_components_radio_additionalLabelText_text")
-        helperTextContent = String(localized: "app_components_radio_helperText_text")
+        labelTextContent = String(localized: "app_components_switch_label_text")
+        helperTextContent = String(localized: "app_components_switch_helperText_text")
     }
 
     deinit { }
@@ -104,34 +89,26 @@ final class RadioItemConfigurationModel: ComponentConfiguration {
     override func updateCode() {
         code =
           """
-        OUDSRadioItem(isOn: $isOn, label: "\(labelTextContent)"\(additionalLabelTextPattern)\(helperTextPattern)\(iconPattern)\(flipIconPattern)\(outlinedPattern)\(isReversedPattern)\(isErrorPattern)\(isReadOnlyPattern)\(dividerPattern))
+        OUDSSwitchItem(isOn: $isOn, label: "\(labelTextContent)"\(helperTextPatern)\(iconPatern)\(flipIconPattern)\(isReversedPattern)\(isErrorPattern)\(isReadOnlyPattern)\(dividerPatern))
         \(disableCode)
         """
     }
     // swiftlint:enable line_length
 
     private var disableCode: String {
-        ".disable(\(enabled ? "false" : "true"))"
+        ".disabled(\(enabled ? "false" : "true"))"
     }
 
-    private var additionalLabelTextPattern: String {
-        helperText ? ", additionalLabel: \"\(additionalLabelTextContent)\"" : ""
-    }
-
-    private var helperTextPattern: String {
+    private var helperTextPatern: String {
         helperText ? ", helper: \"\(helperTextContent)\"" : ""
     }
 
-    private var iconPattern: String {
-        icon ? ", icon: Image(systemName: \"figure.handball\")" : ""
+    private var iconPatern: String {
+        icon ? ", icon: Image(decorative: \"figure.handball\")" : ""
     }
 
     private var flipIconPattern: String {
         flipIcon ? ", flipIcon: true" : ""
-    }
-
-    private var outlinedPattern: String {
-        outlined ? ", isOutlined: true" : ""
     }
 
     private var isReversedPattern: String {
@@ -146,23 +123,22 @@ final class RadioItemConfigurationModel: ComponentConfiguration {
         isReadOnly ? ", isReadOnly: true" : ""
     }
 
-    private var dividerPattern: String {
+    private var dividerPatern: String {
         divider ? ", divider: true" : ""
     }
 }
 
-// MARK: - Radio Item Configuration View
+// MARK: - Switch Item Configuration View
 
-struct RadioItemConfiguration: View {
+struct SwitchItemConfiguration: View {
 
-    @ObservedObject var model: RadioItemConfigurationModel
+    @ObservedObject var model: SwitchItemConfigurationModel
 
     @Environment(\.theme) private var theme
 
-    // swiftlint:disable closure_body_length
     var body: some View {
         VStack(alignment: .leading, spacing: theme.spaces.spaceFixedMedium) {
-            Toggle("app_components_radio_selection_label", isOn: $model.selection)
+            Toggle("app_components_switch_selection_label", isOn: $model.selection)
                 .typeHeadingMedium(theme)
                 .oudsForegroundStyle(theme.colors.colorContentDefault)
                 .disabled(model.isError || model.isReadOnly)
@@ -186,23 +162,17 @@ struct RadioItemConfiguration: View {
                 .typeHeadingMedium(theme)
                 .oudsForegroundStyle(theme.colors.colorContentDefault)
 
-            Toggle("app_components_radio_additionalLabelText_label", isOn: $model.additionalLabelText)
-                .typeHeadingMedium(theme)
-                .oudsForegroundStyle(theme.colors.colorContentDefault)
-
             Toggle("app_components_common_icon_label", isOn: $model.icon)
                 .typeHeadingMedium(theme)
                 .oudsForegroundStyle(theme.colors.colorContentDefault)
 
-            Toggle("app_components_radio_outlined_label", isOn: $model.outlined)
-                .typeHeadingMedium(theme)
-                .oudsForegroundStyle(theme.colors.colorContentDefault)
+            if model.icon {
+                Toggle("app_components_common_flipIcon_label", isOn: $model.flipIcon)
+                    .typeHeadingMedium(theme)
+                    .oudsForegroundStyle(theme.colors.colorContentDefault)
+            }
 
             Toggle("app_components_common_divider_label", isOn: $model.divider)
-                .typeHeadingMedium(theme)
-                .oudsForegroundStyle(theme.colors.colorContentDefault)
-
-            Toggle("app_components_common_flipIcon_label", isOn: $model.flipIcon)
                 .typeHeadingMedium(theme)
                 .oudsForegroundStyle(theme.colors.colorContentDefault)
 
@@ -218,13 +188,6 @@ struct RadioItemConfiguration: View {
                     prompt: "app_components_common_userText_prompt",
                     title: "app_components_common_labelText_label")
 
-                if model.additionalLabelText {
-                    DesignToolboxTextField(
-                        text: $model.additionalLabelTextContent,
-                        prompt: "app_components_common_userText_prompt",
-                        title: "app_components_radio_additionalLabelText_label")
-                }
-
                 if model.helperText {
                     DesignToolboxTextField(
                         text: $model.helperTextContent,
@@ -233,6 +196,5 @@ struct RadioItemConfiguration: View {
                 }
             }
         }
-        // swiftlint:enable closure_body_length
     }
 }