+++ OUDS/Core/Components/Sources/Internal/ControlItem/ControlItem.swift
@@ -108,15 +108,16 @@ struct ControlItem: View {
 
     public var body: some View {
         InteractionButton(isReadOnly: layoutData.isReadOnly) {
-            withAnimation(.easeInOut) {
-                switch indicatorType {
-                case .switch(let binding):
-                    binding.wrappedValue.toggle()
-                case .radioButton(let binding):
-                    binding.wrappedValue.toggle()
-                case .checkBox(let binding):
+            VibrationsManager.success()
+            switch indicatorType {
+            case .switch(let binding):
+                withAnimation(.timingCurve(0.2, 0, 0, 1, duration: 0.150)) {
                     binding.wrappedValue.toggle()
                 }
+            case .radioButton(let binding):
+                binding.wrappedValue.toggle()
+            case .checkBox(let binding):
+                binding.wrappedValue.toggle()
             }
             action?()
         } content: { interactionState in